+++ src/Fable.Core/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* [JS/TS] Fix #3533: Add `emitTopDirectivePrologue` supports (by @MangelMaxime)
-* [JS/TS] Fix #3533: Add `emitDirectivePrologue` supports (by @MangelMaxime)
+* [JS/TS] Fix #3533: Add `emitJsTopDirectivePrologue` supports (by @MangelMaxime)
+* [JS/TS] Fix #3533: Add `emitJsDirectivePrologue` supports (by @MangelMaxime)
 
 ## 4.5.0 - 2025-03-03